Refactor hmac.rb to file-per-method
