+++ google-cloud-storage/samples/acceptance/hmac_test.rb
@@ -13,7 +13,12 @@
 # limitations under the License.
 
 require_relative "helper"
-require_relative "../hmac.rb"
+require_relative "../storage_activate_hmac_key.rb"
+require_relative "../storage_create_hmac_key.rb"
+require_relative "../storage_deactivate_hmac_key.rb"
+require_relative "../storage_delete_hmac_key.rb"
+require_relative "../storage_get_hmac_key.rb"
+require_relative "../storage_list_hmac_keys.rb"
 
 # HMAC key tests are mocked due to the very limited quota of keys per project.
 describe "HMAC Snippets" do
